'Dead-End' Event Processing Block Commented out
In StepZero we find a large code block which at one point likely served the purpose of filtering and separating CSD events from non-CSD events. However, the code block now hits a 'dead-end' and the outputs of the processing block are not used either downstream in the code, or saved as final outputs. Therefore this code block is commented out.