Beginning of Comments for Detection Algorithm
After completing the Preprocessing scripts comments are now being added for the detection algorithm portion of the WAVEFRONT pipeline. This is the portion of the pipeline which actually performs CSD detection after the preprocessing scripts clean and appropriately process the raw data. Because the detection algorithm is much longer and more involved than the preprocessing, the depth and number of comments while be might lighter in the detection script compared to the preprocessing